+++ src/main/java/com/example/onboarding/repository/ReplyRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReplyRepository extends JpaRepository<Reply, Long> {
+public interface ReplyRepository extends JpaRepository<Reply, Integer> {
 
     List<Reply> findAllByGroup_GroupId(int groupGroupId);
-    List<Reply> findByParentReply_ReplyIdOrderByCreatedAtAsc(long parentId);
+    List<Reply> findByParentReply_ReplyIdOrderByCreatedAtAsc(int parentId);
 }
